Delete cruft from array/loader.h and consolidate in arrow/ipc
