+++ cpp/src/arrow/ipc/feather-test.cc
@@ -28,7 +28,6 @@
 #include "arrow/ipc/feather-internal.h"
 #include "arrow/ipc/feather.h"
 #include "arrow/ipc/test-common.h"
-#include "arrow/loader.h"
 #include "arrow/pretty_print.h"
 #include "arrow/test-util.h"
 
@@ -365,11 +364,6 @@ TEST_F(TestTableWriter, TimeTypes) {
   std::shared_ptr<Array> date_array;
   ArrayFromVector<Date32Type, int32_t>(is_valid, date_values_vec, &date_array);
 
-  std::vector<FieldMetadata> fields(1);
-  fields[0].length = values->length();
-  fields[0].null_count = values->null_count();
-  fields[0].offset = 0;
-
   const auto& prim_values = static_cast<const PrimitiveArray&>(*values);
   std::vector<std::shared_ptr<Buffer>> buffers = {
       prim_values.null_bitmap(), prim_values.values()};
@@ -378,9 +372,8 @@ TEST_F(TestTableWriter, TimeTypes) {
   arrays.push_back(date_array->data());
 
   for (int i = 1; i < schema->num_fields(); ++i) {
-    auto arr = std::make_shared<internal::ArrayData>();
-    ASSERT_OK(LoadArray(schema->field(i)->type(), fields, buffers, arr.get()));
-    arrays.emplace_back(arr);
+    arrays.emplace_back(std::make_shared<internal::ArrayData>(
+        schema->field(i)->type(), values->length(), buffers, values->null_count(), 0));
   }
 
   RecordBatch batch(schema, values->length(), std::move(arrays));
